+++ .prettierrc.json
@@ -1,0 +1,9 @@
+{
+    "tabWidth": 2,
+    "useTabs": false,
+    "printWidth": 80,
+    "semi": false,
+    "trailingComma": "es5",
+    "jsxSingleQuote": true,
+    "singleQuote": true
+  }
