--- revent-app/.prettierrc.json
@@ -1,9 +1,0 @@
-{
-    "tabWidth": 2,
-    "useTabs": false,
-    "printWidth": 80,
-    "semi": false,
-    "trailingComma": "es5",
-    "jsxSingleQuote": true,
-    "singleQuote": true
-  }
