--- revent-app/README.md
@@ -1,1 +1,0 @@
-This is a starter template for [Learn Next.js](https://nextjs.org/learn).
