--- revent-app/package.json
@@ -1,21 +1,0 @@
-{
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "lint-staged": "^13.2.2",
-    "next": "latest",
-    "prettier": "^2.8.8",
-    "react": "18.2.0",
-    "react-dom": "18.2.0"
-  },
-  "devDependencies": {
-    "eslint": "8.39.0",
-    "eslint-config-next": "13.3.4",
-    "eslint-config-prettier": "^8.8.0"
-  }
-}
